refactor for torch operator to triton specific
